+++ framework/services/Info.h
@@ -13,12 +13,22 @@ public:
     virtual void onStateChanged(const State& state);
 
 private:
+    String dump() const;
+
     void publish();
+
+#ifdef INFO_HTTP_ENABLED
+    void onHttpIndex(HttpRequest &request, HttpResponse &response);
+#endif
 
     Timer timer;
 
     uint32_t startupTime;
     uint32_t connectTime;
+
+#ifdef INFO_HTTP_ENABLED
+    HttpServer http;
+#endif
 };